+++ core/extensions/Revision.js
@@ -8,19 +8,26 @@ class Revision {
     this.maxUpdates = maxUpdates;
     this.undos = reactive([]);
     this.redos = reactive([]);
+    this.transactionId = null;
 
     this.redoable = computed(() => this.redos.length > 0);
     this.undoable = computed(() => this.undos.length > 0);
   }
 
   // save current state to undo stack
-  commit(patch) {
-    const dontUndo = (val,key) => key.startsWith("settings.")
-    const cleanPatch = patch.map((change) => {
-      return { newValue: omitBy(change.newValue, dontUndo), oldValue: omitBy(change.oldValue, dontUndo) }
-    }).filter((change) => !isEmpty(change.newValue) || !isEmpty(change.oldValue) )
-    if (cleanPatch.length === 0) return
+  commit(patch, transactionId) {
+    const dontUndo = (val, key) => key.startsWith('settings.')
+
+    const cleanPatch = {
+      newValue: omitBy(patch.newValue, dontUndo),
+      oldValue: omitBy(patch.oldValue, dontUndo)
+    };
+
+    if (isEmpty(cleanPatch.newValue) || isEmpty(cleanPatch.oldValue)) return;
+
     this.undos.push(cleanPatch);
+
+    this.transactionId = transactionId;
 
     // Redo is no longer possible once a new change has been made
     this.redos.length = 0;
@@ -35,9 +42,7 @@ class Revision {
     const undoItem = this.undos.pop();
     this.redos.push(undoItem);
 
-    const updates = map(undoItem, 'oldValue');
-
-    this.applyUpdates(updates);
+    this.applyPatch(undoItem.oldValue);
   }
 
   redo() {
@@ -45,25 +50,21 @@ class Revision {
     const redoItem = this.redos.pop();
     this.undos.push(redoItem);
 
-    const updates = map(redoItem, 'newValue').reverse();
-
-    this.applyUpdates(updates);
+    this.applyPatch(redoItem.newValue);
   }
 
-  applyUpdates(updates) {
+  applyPatch(patch) {
     const { transactionDepth, ...state } = this.store.toJSON();
 
-    updates.forEach(patch => {
-      each(patch, (val, key) => {
-        if (val === undefined) {
-          unset(state, key);
-        } else {
-          set(state, key, val);
-        }
-      });
+    each(patch, (val, key) => {
+      if (val === undefined) {
+        unset(state, key);
+      } else {
+        set(state, key, val);
+      }
     });
 
-    this.store.replaceState(state);
+    this.store.replaceState(state, true);
   }
 
   // clear revision undo / redo stack