+++ core/models/ValidationPaneMessage.ts
@@ -1,0 +1,7 @@
+export default interface ValidationPaneMessage {
+    id: string,
+    content: string,
+    timestamp?: number,
+    isDismissable?: boolean,
+    type: 'danger' | 'warning' | 'success' | 'magic' | 'info' | 'default';
+}