+++ core/store/index.js
@@ -1,5 +1,7 @@
+import { nextTick } from 'vue'
 import { defineStore, createPinia } from 'pinia';
 import { normie } from '@crhio/normie';
+import { v4 as uuidv4 } from 'uuid';
 import { useMeta } from '@crhio/leviate';
 import { isEmpty, get, last, range, each, omit, isEqual } from 'lodash-es';
 import Migration from '../extensions/migration';
@@ -10,6 +12,8 @@ import logger from '../extensions/logger.js';
 import useAppInfo from '../composables/useAppInfo.js';
 import useVersions from '../composables/useVersions';
 import { useHost } from '../plugins/host';
+import { useLeviateStore } from './leviate';
+import { checkEntitiesIntegrity } from './storeUtils';
 
 class TransactionError extends Error {
   constructor() {
@@ -23,7 +27,7 @@ let _useStateCompression = false;
 
 let useRootStore = () => logger.log('Root store has not been initialized');
 
-let transactionUpdates = [];
+let transactionUpdates = null;
 
 async function compressState(state) {
   const mod = await import('../composables/useStateCompression.ts');
@@ -37,13 +41,31 @@ async function decompressState(state) {
   return decompress(state);
 }
 
-function transact(cb) {
+async function hostSetState(newState) {
+  const stateToSave = _useStateCompression
+    ? await compressState(newState)
+    : { ...newState, _compressed: undefined };
+
+  const { activeVersionId } = useVersions();
+  return useHost().setState(stateToSave, activeVersionId.value);
+}
+
+function transact(nameOrCallback, callback, options = { skipRevision: false }) {
   if (useMeta().isReadOnly) {
     logger.log('Transaction skipped: Read-only mode.');
     return;
   }
 
-  return useRootStore().transact(cb);
+  let cb = callback;
+  let name = nameOrCallback;
+
+  if (typeof nameOrCallback !== 'string') {
+    logger.warn('legacy transact detected. Transaction will still run but in future you should use `transact(name, callback) for improved logging and debugging`');
+    cb = nameOrCallback;
+    name = 'Anonymous transaction';
+  }
+
+  return useRootStore().transact(name, cb, options);
 }
 
 const initialState = {
@@ -105,58 +127,83 @@ function deepDiff(obj1, obj2) {
 }
 
 const initialActions = {
-  async transact(cb) {
+  async transact(name, cb, options) {
     if (this.transactionDepth === 0) {
-      transactionUpdates = [];
+      transactionUpdates = { oldValue: {}, newValue: {} };
     }
 
     this.transactionDepth++;
 
+    const transactionId = uuidv4();
+    const oldState = this.toJSON();
+
     try {
-      const oldState = this.toJSON();
+      useHost().log?.(`Running transaction "${name}"`);
 
       let res = cb();
 
-      this.transactionDepth --;
+      // Ensure that any nested transactions within reactive handlers
+      // are processed before continuing
+      await nextTick();
 
       if (res instanceof Promise) {
-        await res.catch(this._onTransactionError);
+        await res.catch((e) => {
+          throw e;
+        });
       }
 
       const newState = this.toJSON();
 
       const diff = deepDiff(oldState, newState);
 
-      const stateToSave = _useStateCompression
-        ? await compressState(newState)
-        : { ...diff.newValue, _compressed: undefined };
+      const stateToSave = _useStateCompression ? newState : diff.newValue;
+      const updateKeys = { keys: Object.keys(stateToSave) };
 
-      const { activeVersionId } = useVersions();
-      useHost().setState(stateToSave, activeVersionId.value);
-      transactionUpdates.unshift(diff);
-
-      if (this.transactionDepth === 0) {
-        this.revision.commit(transactionUpdates);
+      // Warn if there is nothing to update
+      if (updateKeys.keys.length === 0) {
+        logger.warn(`Nothing to save in transaction ${name}`);
       }
+
+      if (!options.skipRevision) {
+        Object.assign(transactionUpdates.oldValue, diff.oldValue);
+        Object.assign(transactionUpdates.newValue, diff.newValue);
+      }
+
+      hostSetState(stateToSave);
+
+      this.transactionDepth --;
+
+      if (this.transactionDepth === 0 && !isEmpty(transactionUpdates.newValue)) {
+        this.revision.commit(transactionUpdates, transactionId);
+      }
+
+      useHost().log?.(`Transaction "${name}" completed successfully`, updateKeys);
 
       return res;
     } catch (e) {
-      this._onTransactionError(e);
+      this._onTransactionError(e, name, transactionId, oldState);
       return false;
     }
   },
 
-  _onTransactionError(e) {
+  _onTransactionError(e, name, transactionId, oldState) {
     if (!(e instanceof TransactionError)) {
-      logger.error(e);
+      useHost().log?.(`Transaction "${name}" failed`);
+      logger.error('transaction failed -', e);
     }
 
     if (this.transactionDepth > 1) {
       // if we're in a nested transaction, propagate an error;
       throw new TransactionError();
     } else {
-      // otherwise, undo to the last committed state;
-      this.revision.undo();
+      if (transactionId === this.revision.transactionId) {
+        this.revision.undo();
+      } else {
+        this.replaceState(oldState, true);
+      }
+
+      useLeviateStore().setGlobalMessage(useLocalize().$L('error_global'));
+
       this.transactionDepth = 0;
     }
   },
@@ -176,10 +223,11 @@ const initialActions = {
     this.modules[useStore.$id] = useStore;
   },
 
-  replaceState(newState) {
+  replaceState(newState, shouldSync = false) {
+    const serializedState = JSON.parse(JSON.stringify(newState));
     const rootState = {};
 
-    Object.entries(newState).forEach(([key, value]) => {
+    Object.entries(serializedState).forEach(([key, value]) => {
       if (this.$state[key] !== undefined) {
         rootState[key] = value;
         return;
@@ -188,13 +236,17 @@ const initialActions = {
       const useModule = this.modules[key];
       if (useModule) {
         const module = useModule();
-          module.$state = newState[key];
+        module.$state = serializedState[key];
       } else {
         logger.error(`Store module ${key} does not exist`);
       }
     });
 
     this.$patch(rootState);
+
+    if (shouldSync) {
+      hostSetState(serializedState);
+    }
   },
 
   toJSON() {
@@ -358,9 +410,15 @@ function performMigration(rootStore, initialState, migrations) {
       migratedState = migration.migrateToLatest();
       logger.log(`successfully migrated to latest: ${latestMigrationName}`);
     }
-    rootStore.replaceState(migratedState || initialState);
+    rootStore.replaceState(migratedState || initialState, true);
   } else {
-    rootStore.$patch({ serialization_version: latestMigrationName });
+    hostSetState({ serialization_version: latestMigrationName });
+  }
+
+  const modifiedEntities = checkEntitiesIntegrity(rootStore);
+
+  if (modifiedEntities) {
+    hostSetState({ entities: modifiedEntities });
   }
 }